requirements.txt
python-dotenv
youtube_transcript_api
openai
streamlit
google-api-python-client
# supabase-py
